fix Third version review issues

+++ contracts/feeDistributorFactory/FeeDistributorFactory.sol
@@ -60,7 +60,7 @@ contract FeeDistributorFactory is OwnableAssetRecoverer, OwnableWithOperator, ER
     * @param _client the address of the client
     * @param _serviceBasisPoints basis points (percent * 100) of EL rewards that should go to the service (P2P)
     */
-    function createFeeDistributor(address _client, uint256 _serviceBasisPoints) external onlyOperatorOrOwner {
+    function createFeeDistributor(address _client, uint96 _serviceBasisPoints) external onlyOperatorOrOwner {
         if (s_referenceFeeDistributor == address(0)) {
             revert FeeDistributorFactory__ReferenceFeeDistributorNotSet();
         }
